requirements.txt
alembic==1.11.2
asyncpg==0.28.0
envparse==0.2.0
fastapi==0.101.1
psycopg==3.1.10
pydantic[email]~=2.3.0
SQLAlchemy==2.0.19
uvicorn==0.23.2
pytest~=7.4.0
httpx
starlette~=0.27.0
python-multipart
passlib~=1.7.4
pre-commit
python-jose
